Add void* argument to add_point function typedef and the implementations.
Initially, this is to permit a test with using stack memory for
the linked list (using alloca() in the loop for the memory.  It can
be ignored by an implementation, or used as an integer value to
reference other data structures.

The stack memory test will be done to see if the time required to
allocate stack memory is more consistent than heap allocation that
needs additional time at predictable intervals (at a memory page?).

Using stack memory in this way is inherently unsafe because the
object to which the stack memory will outlive the scope of the stack
memory.  If I decide that using stack memory is the preferred
method, I will have to devise a way to improve the safety of its
use.